TimeUtil 클래스 추가 -> ?시간전 기능 모듈화
